backend/pyproject.toml
[tool.poetry]
name = "arema-api"
version = "0.1.0"
description = ""
authors = ["ymarcon <yannick.marcon@epfl.ch>"]
readme = "README.md"
packages = [{include = "api"}]
include = ["api/data/*.yaml"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.4"
enacit4r-files = {git = "https://github.com/EPFL-ENAC/enacit4r-files", rev = "0.4.0"}
enacit4r-auth = {git = "https://github.com/EPFL-ENAC/enacit4r-auth", rev = "0.1.0"}
enacit4r-sql = {git = "https://github.com/EPFL-ENAC/enacit4r-sql", rev = "0.1.0"}
alembic = "^1.14.0"
asyncpg = "^0.30.0"
psycopg2 = "^2.9.10"
pydantic-settings = "^2.6.1"
dynaconf = "^3.2.6"
python-multipart = "^0.0.17"
python-dotenv = {extras = ["cli"], version = "^1.0.1"}
uvicorn = "^0.32.0"
pyyaml = "^6.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
